Move `packwerk update` logic to Cli::UpdateCommand
